pyproject.toml
[tool.poetry]
name = "fastapi-lead-quote-api"
version = "0.1.0"
description = "A FastAPI project for Lead Transfer and Quick Quote APIs"
authors = ["Your Name <youremail@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.0"  # Updated for Pydantic 2.x compatibility
uvicorn = "^0.23.0"
sqlmodel = "^0.0.8"
asyncpg = "^0.28.0"
pydantic = "^2.4.0"  # Updated to latest Pydantic
pydantic-settings = "^2.0.0"  # Added for settings management
alembic = "^1.12.0"
pytest = "^7.4.0"
httpx = "^0.25.0"
python-dotenv = "^1.0.0"
email-validator = "^2.0.0"  # Required for EmailStr
passlib = "^1.7.4"
python-multipart = "^0.0.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"